+++ CHANGELOG.md
@@ -30,5 +30,6 @@
 * **features:** Add comparison box list to Activity patterns page (#129)
 * **feature:** Table pagination (#135)
 * **feature:** Update the table row to be consistent row (#160)
+* **feature:** Table sort by species name or class (#147)
 
 * **refactor:** Refactor species richness table display (#112)